DALib cycle 2: MD5 Encode/Decode endian helpers (correct partials 75%/70%)
RSA-reference MD5 Encode/Decode (little-endian dword<->byte). KEY: the MD5 dword
type is unsigned long (K mangling) not unsigned int (I) -- got 0% blank pairing
until fixed. Correct logic; residual is the original's byte-copy optimization of
the shifts, not reproducible from MSVC6 here.

+++ src/DALib.h
@@ -20,8 +20,8 @@ public:
 
 protected:
     void Transform(unsigned char* block);
-    void Encode(unsigned char* output, unsigned int* input, unsigned int len);
-    void Decode(unsigned int* output, unsigned char* input, unsigned int len);
+    void Encode(unsigned char* output, unsigned long* input, unsigned long len);
+    void Decode(unsigned long* output, unsigned char* input, unsigned long len);
 };
 
 // --- Win32 critical section (minimal; avoids pulling in <windows.h>) ---
